Profile fields

+++ application/modules/user/mappers/User.php
@@ -185,38 +185,6 @@ class User extends \Ilch\Mapper
             $user->setGender($userRow['gender']);
         }
 
-        if (isset($userRow['homepage'])) {
-            $user->setHomepage($userRow['homepage']);
-        }
-
-        if (isset($userRow['facebook'])) {
-            $user->setFacebook($userRow['facebook']);
-        }
-
-        if (isset($userRow['twitter'])) {
-            $user->setTwitter($userRow['twitter']);
-        }
-
-        if (isset($userRow['google'])) {
-            $user->setGoogle($userRow['google']);
-        }
-
-        if (isset($userRow['steam'])) {
-            $user->setSteam($userRow['steam']);
-        }
-
-        if (isset($userRow['twitch'])) {
-            $user->setTwitch($userRow['twitch']);
-        }
-
-        if (isset($userRow['teamspeak'])) {
-            $user->setTeamspeak($userRow['teamspeak']);
-        }
-
-        if (isset($userRow['discord'])) {
-            $user->setDiscord($userRow['discord']);
-        }
-
         if (isset($userRow['city'])) {
             $user->setCity($userRow['city']);
         }
@@ -350,14 +318,6 @@ class User extends \Ilch\Mapper
         $fields['first_name'] = $user->getFirstName();
         $fields['last_name'] = $user->getLastName();
         $fields['gender'] = $user->getGender();
-        $fields['homepage'] = $user->getHomepage();
-        $fields['facebook'] = $user->getFacebook();
-        $fields['twitter'] = $user->getTwitter();
-        $fields['google'] = $user->getGoogle();
-        $fields['steam'] = $user->getSteam();
-        $fields['twitch'] = $user->getTwitch();
-        $fields['teamspeak'] = $user->getTeamspeak();
-        $fields['discord'] = $user->getDiscord();
         $fields['city'] = $user->getCity();
         $fields['birthday'] = $user->getBirthday();
         $fields['avatar'] = $user->getAvatar();
@@ -492,6 +452,10 @@ class User extends \Ilch\Mapper
         }
 
         $this->db()->delete('users_groups')
+            ->where(['user_id' => $userId])
+            ->execute();
+
+        $this->db()->delete('profile_content')
             ->where(['user_id' => $userId])
             ->execute();
